+++ algg.c
@@ -188,12 +188,12 @@ static void enqueueFinAC(cl_command_queue q, size_t height, size_t k,
 			 size_t skip, cl_mem from, cl_mem to, size_t n) {
   size_t ori[3] = {0, 0, 0};
   size_t foo = k - skip;
-  size_t reg[3] = {sizeof(double), foo, n};
+  size_t reg[3] = {sizeof(float), foo, n};
   if(clEnqueueCopyBufferRect(q, from, to, ori, ori, reg,
-			     height * sizeof(double),
-			     height * sizeof(double) * foo,
-			     sizeof(double),
-			     sizeof(double) * foo,
+			     height * sizeof(float),
+			     height * sizeof(float) * foo,
+			     sizeof(float),
+			     sizeof(float) * foo,
 			     0, NULL, NULL) != CL_SUCCESS)
     fprintf(stderr, "Failed enqueue of copy.\n"), exit(1);
 }
