+++ rand_pr.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 extern void rand_rot(size_t rotlen, size_t d,
-		     size_t *is, size_t *js, double *as);
+		     size_t *is, size_t *js, float *as);
 extern size_t *rand_perm(size_t d_pre, size_t d_post);
 
 #endif
